+++ packages/workbench/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @computesdk/workbench
 
+## 3.1.3
+
+### Patch Changes
+
+- Updated dependencies [acdc8c6]
+  - @computesdk/provider@1.0.3
+  - @computesdk/e2b@1.7.9
+  - @computesdk/modal@1.8.4
+  - @computesdk/cloudflare@1.3.9
+  - @computesdk/blaxel@1.3.9
+  - @computesdk/codesandbox@1.5.9
+  - @computesdk/daytona@1.6.9
+  - @computesdk/vercel@1.6.9
+  - @computesdk/runloop@1.3.9
+  - @computesdk/railway@1.1.9
+
 ## 3.1.2
 
 ### Patch Changes